backend/requirements.txt
fastapi[all]==0.109.2
gunicorn==21.2.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-cloud-storage==2.16.0
